設定画面の見た目を改善(ミュート管理画面をFragmentで遷移するように)
