+++ app/src/main/res/xml/preferences_restrict.xml
@@ -1,0 +1,18 @@
+<?xml version="1.0" encoding="utf-8"?>
+<PreferenceScreen xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <PreferenceCategory android:title="@string/pref_manage_restrictions">
+
+        <PreferenceScreen android:key="manage_mute_accounts"
+            android:title="@string/pref_manage_mute_accounts" />
+
+        <PreferenceScreen android:key="manage_mute_keywords"
+            android:title="@string/pref_manage_mute_keywords" />
+
+        <PreferenceScreen android:key="manage_mute_hash_tags"
+            android:title="@string/pref_manage_mute_hash_tags" />
+
+        <PreferenceScreen android:key="manage_mute_instances"
+            android:title="@string/pref_manage_mute_instances" />
+    </PreferenceCategory>
+</PreferenceScreen>
